--- reasoning_diff_project/non_chat_template_tokenizer_analysis.txt
@@ -1,77 +1,0 @@
-Loading models...
-Loading dataset...
-
-Running tokenizer compatibility tests...
-
-Base with base...
-Example No. 0
-Please reason step by step, and put your final answer within \boxed{}.3. As shown in the figure, a toy ant starts crawling from point $O$. On the $n$-th time, it first crawls $n$ units to the right, then $n$ units upwards, reaching point $A_{n}$. Then, it continues crawling from point $A_{n}$. If point $O$ is denoted as $(0,0)$, point $A_{1}$ as $(1,1)$, point $A_{2}$ as $(3,3)$, point $A_{3}$ as $(6,6), \cdots$, then point $A_{100}$ is denoted as $\qquad$.
-------------------------------------------------------
-Please reason step by step, and put your final answer within \boxed{}.3. As shown in the figure, a toy ant starts crawling from point $O$. On the $n$-th time, it first crawls $n$ units to the right, then $n$ units upwards, reaching point $A_{n}$. Then, it continues crawling from point $A_{n}$. If point $O$ is denoted as $(0,0)$, point $A_{1}$ as $(1,1)$, point $A_{2}$ as $(3,3)$, point $A_{3}$ as $(6,6), \cdots$, then point $A_{100}$ is denoted as $\qquad$. To determine the coordinates of point \(A_{100}\), we need to analyze the pattern of the ant's movement. The ant starts at the origin \((0,0)\) and moves \(n\) units to the right and then \(n\) units upwards on the \(n\)-th move. Therefore, the coordinates of point \(A_n\) can be determined by summing the movements in the x and y directions.
-
-The x-coordinate of \(A_n\) is the sum of the first \(n\) natural numbers, and the y-coordinate of \(A_n\) is also the sum of the first \(n\) natural numbers. The sum of the first \(n\) natural numbers is given by the formula:
-\[
-S_n = \frac{n(n+1)}{2}
-\]
-So, the coordinates of point \(A_{100}\) will be:
-\[
-A_{100} = \left( \frac{100 \cdot 101}{2}, \frac{100 \cdot 101}{2} \right) = (5050, 5050)
-\]
-
-Let's confirm this with Python code.
-```python
-# Calculate the coordinates of point A_100
-n = 100
-x_coordinate = n * (n + 1) // 2
-y_coordinate = n * (n + 1) // 2
-
-print(((x_coordinate, y_coordinate)))
-```
-```output
-(5050, 5050)
-```
-The coordinates of point \(A_{100}\) are \(\boxed{(5050, 5050)}\).
-======================================================
-
-Base with ft...
-Example No. 0
-Please reason step by step, and put your final answer within \boxed{}.3. As shown in the figure, a toy ant starts crawling from point $O$. On the $n$-th time, it first crawls $n$ units to the right, then $n$ units upwards, reaching point $A_{n}$. Then, it continues crawling from point $A_{n}$. If point $O$ is denoted as $(0,0)$, point $A_{1}$ as $(1,1)$, point $A_{2}$ as $(3,3)$, point $A_{3}$ as $(6,6), \cdots$, then point $A_{100}$ is denoted as $\qquad$.
-------------------------------------------------------
-Please reason step by step, and put your final answer within \boxed{}.3. As shown in the figure, a toy ant starts crawling from point $O$. On the $n$-th time, it first crawls $n$ units to the right, then $n$ units upwards, reaching point $A_{n}$. Then, it continues crawling from point $A_{n}$. If point $O$ is denoted as $(0,0)$, point $A_{1}$ as $(1,1)$, point $A_{2}$ as $(3,3)$, point $A_{3}$ as $(6,6), \cdots$, then point $A_{100}$ is denoted as $\qquad$. To determine the coordinates of point \(A_{100}\), we need to analyze the pattern of the ant's movement. The ant starts at the origin \((0,0)\) and moves \(n\) units to the right and then \(n\) units upwards on the \(n\)-th move. Therefore, the coordinates of point \(A_n\) can be determined by summing the movements in the x and y directions.
-
-The x-coordinate of \(A_n\) is the sum of the first \(n\) natural numbers, and the y-coordinate of \(A_n\) is also the sum of the first \(n\) natural numbers. The sum of the first \(n\) natural numbers is given by the formula:
-\[
-S_n = \frac{n(n+1)}{2}
-\]
-So, the coordinates of point \(A_{100}\) will be:
-\[
-A_{100} = \left( \frac{100 \cdot 101}{2}, \frac{100 \cdot 101}{2} \right) = (5050, 5050)
-\]
-
-Let's confirm this with Python code.
-```python
-# Calculate the coordinates of point A_100
-n = 100
-x_coordinate = n * (n + 1) // 2
-y_coordinate = n * (n + 1) // 2
-
-print(((x_coordinate, y_coordinate)))
-```
-```output
-(5050, 5050)
-```
-The coordinates of point \(A_{100}\) are \(\boxed{(5050, 5050)}\).
-======================================================
-\Ft with ft...
-Example No. 0
-Please reason step by step, and put your final answer within \boxed{}.3. As shown in the figure, a toy ant starts crawling from point $O$. On the $n$-th time, it first crawls $n$ units to the right, then $n$ units upwards, reaching point $A_{n}$. Then, it continues crawling from point $A_{n}$. If point $O$ is denoted as $(0,0)$, point $A_{1}$ as $(1,1)$, point $A_{2}$ as $(3,3)$, point $A_{3}$ as $(6,6), \cdots$, then point $A_{100}$ is denoted as $\qquad$.
-------------------------------------------------------
-Please reason step by step, and put your final answer within \boxed{}.3. As shown in the figure, a toy ant starts crawling from point $O$. On the $n$-th time, it first crawls $n$ units to the right, then $n$ units upwards, reaching point $A_{n}$. Then, it continues crawling from point $A_{n}$. If point $O$ is denoted as $(0,0)$, point $A_{1}$ as $(1,1)$, point $A_{2}$ as $(3,3)$, point $A_{3}$ as $(6,6), \cdots$, then point $A_{100}$ is denoted as $\qquad$. Then, point $A_{100}$ is denoted as $\qquad. Then,
-======================================================
-\Ft with Base...
-Example No. 0
-Please reason step by step, and put your final answer within \boxed{}.3. As shown in the figure, a toy ant starts crawling from point $O$. On the $n$-th time, it first crawls $n$ units to the right, then $n$ units upwards, reaching point $A_{n}$. Then, it continues crawling from point $A_{n}$. If point $O$ is denoted as $(0,0)$, point $A_{1}$ as $(1,1)$, point $A_{2}$ as $(3,3)$, point $A_{3}$ as $(6,6), \cdots$, then point $A_{100}$ is denoted as $\qquad$.
-------------------------------------------------------
-Please reason step by step, and put your final answer within \boxed{}.3. As shown in the figure, a toy ant starts crawling from point $O$. On the $n$-th time, it first crawls $n$ units to the right, then $n$ units upwards, reaching point $A_{n}$. Then, it continues crawling from point $A_{n}$. If point $O$ is denoted as $(0,0)$, point $A_{1}$ as $(1,1)$, point $A_{2}$ as $(3,3)$, point $A_{3}$ as $(6,6), \cdots$, then point $A_{100}$ is denoted as $\qquad$.  
-Okay, so I need to figure out the coordinates of point A₁₀₀. The problem
-======================================================
